feat: added cgaz tool to geohub
