+++ sites/geohub/package.json
@@ -52,6 +52,7 @@
 		"@types/uuid": "^9.0.3",
 		"@typescript-eslint/eslint-plugin": "6.3.0",
 		"@typescript-eslint/parser": "^6.5.0",
+		"@undp-data/cgaz-admin-tool": "workspace:^",
 		"@undp-data/current-location": "workspace:*",
 		"@undp-data/style-switcher": "workspace:*",
 		"@undp-data/svelte-undp-design": "workspace:*",
