+++ sites/geohub/src/lib/config/AppConfig/index.ts
@@ -1,5 +1,6 @@
 export * from './AcceptedExtensions';
 export * from './AccessLevel';
+export * from './AdminControlOptions';
 export * from './attribution';
 export * from './ClassificationMethod';
 export * from './ClassificationMethods';
